Added basic README and fixed PackageRepo datatype

+++ src/main/java/org/group4/travelexpertsapi/repository/PackageRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PackageRepo extends JpaRepository<Package, Long> {
+public interface PackageRepo extends JpaRepository<Package, Integer> {
 
 }
